breaking change, removed lots of unused files in py and lots of file structure change in js
