+++ ui/src/RecentFilesDrawer/AddTagToWorkflowPopover.tsx
@@ -12,7 +12,7 @@ import {
   IconButton,
 } from "@chakra-ui/react";
 import { useEffect, useState, useContext } from "react";
-import { Tag, Workflow, tagsTable, updateFlow } from "../WorkspaceDB";
+import { Tag, Workflow, tagsTable, updateFlow } from "../db-tables/WorkspaceDB";
 import { IconPlus, IconTag } from "@tabler/icons-react";
 import { MultiValue, Select } from "chakra-react-select";
 import { RecentFilesContext } from "../WorkspaceContext";
