+++ ui/src/db-tables/indexdb.ts
@@ -1,17 +1,24 @@
 // db.ts
 import Dexie, { Table } from "dexie";
-import { Workflow } from "../WorkspaceDB";
+import { Workflow } from "./WorkspaceDB";
+import { Changelog, Media } from "../types/dbTypes";
 
-export class WorkspaceDB extends Dexie {
-  // 'friends' is added by dexie when declaring the stores()
-  // We just tell the typing system this is the case
+class WorkspaceDB extends Dexie {
   workflows!: Table<Workflow, string>;
+  changelogs!: Table<Changelog, string>;
+  media!: Table<Media, string>;
 
   constructor() {
     super("WorkspaceDB");
-    this.version(1).stores({
-      workflows: "++id, name", // Primary key and indexed props
-    });
+    this.version(1)
+      .stores({
+        workflows: "++id, name", // Primary key and indexed props
+        changelogs: "++id, workflowID",
+        media: "++id, workflowID",
+      })
+      .upgrade((trans) => {
+        // Here you can write logic to initialize or migrate data to the new 'media' table, if necessary
+      });
   }
 }
