+++ ui/src/utils.tsx
@@ -2,13 +2,13 @@
 import { deleteFile } from "./Api";
 import { ESortTypes } from "./RecentFilesDrawer/types";
 import {
-  Folder,
   listWorkflows,
   saveJsonFileMyWorkflows,
   userSettingsTable,
   Workflow,
-} from "./WorkspaceDB";
+} from "./db-tables/WorkspaceDB";
 import { generateFilePathAbsolute } from "./db-tables/DiskFileUtils";
+import { Folder } from "./types/dbTypes";
 // @ts-ignore
 import { app, ComfyApp } from "/scripts/app.js";
